# docs/openapi-user.json
{"openapi":"3.0.1","info":{"title":"苍穹外卖项目接口文档","description":"苍穹外卖项目接口文档","version":"2.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"userToken":[]}],"paths":{"/user/order/payment":{"put":{"tags":["C端-订单接口"],"summary":"订单支付","operationId":"payment","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrdersPaymentDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultOrderPaymentVO"}}}}}}},"/user/order/cancel/{id}":{"put":{"tags":["C端-订单接口"],"summary":"用户取消订单","operationId":"cancel","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}}}}},"/user/addressBook":{"put":{"tags":["C端地址簿接口"],"summary":"根据id修改地址","operationId":"update","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddressBook"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}}}},"post":{"tags":["C端地址簿接口"],"summary":"新增地址","operationId":"save","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddressBook"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}}}},"delete":{"tags":["C端地址簿接口"],"summary":"根据id删除地址","operationId":"deleteById","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}}}}},"/user/addressBook/default":{"get":{"tags":["C端地址簿接口"],"summary":"查询默认地址","operationId":"getDefault","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultAddressBook"}}}}}},"put":{"tags":["C端地址簿接口"],"summary":"设置默认地址","operationId":"setDefault","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddressBook"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}}}}},"/user/user/login":{"post":{"tags":["C端用户相关接口"],"summary":"微信登录","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultUserLoginVO"}}}}}}},"/user/shoppingCart/sub":{"post":{"tags":["C端-购物车接口"],"summary":"删除购物车中一个商品","operationId":"sub","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShoppingCartDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}}}}},"/user/shoppingCart/add":{"post":{"tags":["C端-购物车接口"],"summary":"添加购物车","operationId":"add","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShoppingCartDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultString"}}}}}}},"/user/order/submit":{"post":{"tags":["C端-订单接口"],"summary":"用户下单","operationId":"submit","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrdersSubmitDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultOrderSubmitVO"}}}}}}},"/user/order/repetition/{id}":{"post":{"tags":["C端-订单接口"],"summary":"再来一单","operationId":"repetition","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}}}}},"/user/shoppingCart/list":{"get":{"tags":["C端-购物车接口"],"summary":"查看购物车","operationId":"list","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultListShoppingCart"}}}}}}},"/user/shop/status":{"get":{"tags":["店铺相关接口"],"summary":"获取店铺的营业状态","operationId":"getStatus","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultInteger"}}}}}}},"/user/setmeal/list":{"get":{"tags":["C端-套餐浏览接口"],"summary":"根据分类id查询套餐","operationId":"list_1","parameters":[{"name":"categoryId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultListSetmeal"}}}}}}},"/user/setmeal/dish/{id}":{"get":{"tags":["C端-套餐浏览接口"],"summary":"根据套餐id查询包含的菜品列表","operationId":"dishList","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultListDishItemVO"}}}}}}},"/user/order/reminder/{id}":{"get":{"tags":["C端-订单接口"],"summary":"用户催单","operationId":"reminder","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}}}}},"/user/order/orderDetail/{id}":{"get":{"tags":["C端-订单接口"],"summary":"查询订单详情","operationId":"details","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultOrderVO"}}}}}}},"/user/order/historyOrders":{"get":{"tags":["C端-订单接口"],"summary":"历史订单查询","operationId":"page","parameters":[{"name":"pageNum","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"pageSize","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"status","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultPageResult"}}}}}}},"/user/dish/list":{"get":{"tags":["C端-菜品浏览接口"],"summary":"根据分类id查询菜品","operationId":"list_2","parameters":[{"name":"categoryId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultListDishVO"}}}}}}},"/user/category/list":{"get":{"tags":["C端-分类接口"],"summary":"查询分类","operationId":"list_3","parameters":[{"name":"type","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultListCategory"}}}}}}},"/user/addressBook/{id}":{"get":{"tags":["C端地址簿接口"],"summary":"根据id查询地址","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultAddressBook"}}}}}}},"/user/addressBook/list":{"get":{"tags":["C端地址簿接口"],"summary":"查询当前登录用户的所有地址信息","operationId":"list_4","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultListAddressBook"}}}}}}},"/user/shoppingCart/clean":{"delete":{"tags":["C端-购物车接口"],"summary":"清空购物车商品","operationId":"clean","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultString"}}}}}}}},"components":{"schemas":{"OrdersPaymentDTO":{"type":"object","properties":{"orderNumber":{"type":"string","description":"订单号"},"payMethod":{"type":"integer","description":"付款方式","format":"int32"}},"description":"订单支付传递的数据模型"},"OrderPaymentVO":{"type":"object","properties":{"nonceStr":{"type":"string","description":"随机字符串"},"paySign":{"type":"string","description":"签名"},"timeStamp":{"type":"string","description":"时间戳"},"signType":{"type":"string","description":"签名算法"},"packageStr":{"type":"string","description":"统一下单接口返回的 prepay_id 参数值"}},"description":"订单支付返回的数据格式"},"ResultOrderPaymentVO":{"type":"object","properties":{"code":{"type":"integer","description":"编码：1成功，0和其它数字为失败","format":"int32"},"msg":{"type":"string","description":"错误信息"},"data":{"$ref":"#/components/schemas/OrderPaymentVO"}},"description":"后端统一返回结果"},"Result":{"type":"object","properties":{"code":{"type":"integer","description":"编码：1成功，0和其它数字为失败","format":"int32"},"msg":{"type":"string","description":"错误信息"},"data":{"type":"object","description":"数据"}},"description":"后端统一返回结果"},"AddressBook":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int64"},"consignee":{"type":"string"},"phone":{"type":"string"},"sex":{"type":"string"},"provinceCode":{"type":"string"},"provinceName":{"type":"string"},"cityCode":{"type":"string"},"cityName":{"type":"string"},"districtCode":{"type":"string"},"districtName":{"type":"string"},"detail":{"type":"string"},"label":{"type":"string"},"isDefault":{"type":"integer","format":"int32"}}},"UserLoginDTO":{"type":"object","properties":{"code":{"type":"string","description":"微信登录授权码"}},"description":"C端用户登录时传递的数据模型"},"ResultUserLoginVO":{"type":"object","properties":{"code":{"type":"integer","description":"编码：1成功，0和其它数字为失败","format":"int32"},"msg":{"type":"string","description":"错误信息"},"data":{"$ref":"#/components/schemas/UserLoginVO"}},"description":"后端统一返回结果"},"UserLoginVO":{"type":"object","properties":{"id":{"type":"integer","description":"主键值","format":"int64"},"openid":{"type":"string","description":"微信用户唯一标识"},"token":{"type":"string","description":"jwt令牌"}},"description":"C端用户登录返回的数据格式"},"ShoppingCartDTO":{"type":"object","properties":{"dishId":{"type":"integer","description":"菜品id","format":"int64"},"setmealId":{"type":"integer","description":"套餐id","format":"int64"},"dishFlavor":{"type":"string","description":"菜品口味"}},"description":"购物车操作传递的数据模型"},"ResultString":{"type":"object","properties":{"code":{"type":"integer","description":"编码：1成功，0和其它数字为失败","format":"int32"},"msg":{"type":"string","description":"错误信息"},"data":{"type":"string","description":"数据"}},"description":"后端统一返回结果"},"OrdersSubmitDTO":{"type":"object","properties":{"addressBookId":{"type":"integer","description":"地址簿id","format":"int64"},"payMethod":{"type":"integer","description":"付款方式","format":"int32"},"remark":{"type":"string","description":"备注"},"estimatedDeliveryTime":{"type":"string","description":"预计送达时间","format":"date-time"},"deliveryStatus":{"type":"integer","description":"配送状态 1立即送出 0选择具体时间","format":"int32"},"tablewareNumber":{"type":"integer","description":"餐具数量","format":"int32"},"tablewareStatus":{"type":"integer","description":"餐具数量状态 1按餐量提供 0选择具体数量","format":"int32"},"packAmount":{"type":"integer","description":"打包费","format":"int32"},"amount":{"type":"number","description":"总金额"}},"description":"用户下单传递的数据模型"},"OrderSubmitVO":{"type":"object","properties":{"id":{"type":"integer","description":"订单id","format":"int64"},"orderNumber":{"type":"string","description":"订单号"},"orderAmount":{"type":"number","description":"订单金额"},"orderTime":{"type":"string","description":"下单时间","format":"date-time"}},"description":"用户下单返回的数据格式"},"ResultOrderSubmitVO":{"type":"object","properties":{"code":{"type":"integer","description":"编码：1成功，0和其它数字为失败","format":"int32"},"msg":{"type":"string","description":"错误信息"},"data":{"$ref":"#/components/schemas/OrderSubmitVO"}},"description":"后端统一返回结果"},"ResultListShoppingCart":{"type":"object","properties":{"code":{"type":"integer","description":"编码：1成功，0和其它数字为失败","format":"int32"},"msg":{"type":"string","description":"错误信息"},"data":{"type":"array","description":"数据","items":{"$ref":"#/components/schemas/ShoppingCart"}}},"description":"后端统一返回结果"},"ShoppingCart":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"userId":{"type":"integer","format":"int64"},"dishId":{"type":"integer","format":"int64"},"setmealId":{"type":"integer","format":"int64"},"dishFlavor":{"type":"string"},"number":{"type":"integer","format":"int32"},"amount":{"type":"number"},"image":{"type":"string"},"createTime":{"type":"string","format":"date-time"}},"description":"数据"},"ResultInteger":{"type":"object","properties":{"code":{"type":"integer","description":"编码：1成功，0和其它数字为失败","format":"int32"},"msg":{"type":"string","description":"错误信息"},"data":{"type":"integer","description":"数据","format":"int32"}},"description":"后端统一返回结果"},"ResultListSetmeal":{"type":"object","properties":{"code":{"type":"integer","description":"编码：1成功，0和其它数字为失败","format":"int32"},"msg":{"type":"string","description":"错误信息"},"data":{"type":"array","description":"数据","items":{"$ref":"#/components/schemas/Setmeal"}}},"description":"后端统一返回结果"},"Setmeal":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"categoryId":{"type":"integer","format":"int64"},"name":{"type":"string"},"price":{"type":"number"},"status":{"type":"integer","format":"int32"},"description":{"type":"string"},"image":{"type":"string"},"createTime":{"type":"string","format":"date-time"},"updateTime":{"type":"string","format":"date-time"},"createUser":{"type":"integer","format":"int64"},"updateUser":{"type":"integer","format":"int64"}},"description":"数据"},"DishItemVO":{"type":"object","properties":{"name":{"type":"string","description":"菜品名称"},"copies":{"type":"integer","description":"份数","format":"int32"},"image":{"type":"string","description":"菜品图片"},"description":{"type":"string","description":"菜品描述"}},"description":"套餐内菜品项返回的数据格式"},"ResultListDishItemVO":{"type":"object","properties":{"code":{"type":"integer","description":"编码：1成功，0和其它数字为失败","format":"int32"},"msg":{"type":"string","description":"错误信息"},"data":{"type":"array","description":"数据","items":{"$ref":"#/components/schemas/DishItemVO"}}},"description":"后端统一返回结果"},"OrderDetail":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"orderId":{"type":"integer","format":"int64"},"dishId":{"type":"integer","format":"int64"},"setmealId":{"type":"integer","format":"int64"},"dishFlavor":{"type":"string"},"number":{"type":"integer","format":"int32"},"amount":{"type":"number"},"image":{"type":"string"}},"description":"订单详情"},"OrderVO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"number":{"type":"string"},"status":{"type":"integer","format":"int32"},"userId":{"type":"integer","format":"int64"},"addressBookId":{"type":"integer","format":"int64"},"orderTime":{"type":"string","format":"date-time"},"checkoutTime":{"type":"string","format":"date-time"},"payMethod":{"type":"integer","format":"int32"},"payStatus":{"type":"integer","format":"int32"},"amount":{"type":"number"},"remark":{"type":"string"},"userName":{"type":"string"},"phone":{"type":"string"},"address":{"type":"string"},"consignee":{"type":"string"},"cancelReason":{"type":"string"},"rejectionReason":{"type":"string"},"cancelTime":{"type":"string","format":"date-time"},"estimatedDeliveryTime":{"type":"string","format":"date-time"},"deliveryStatus":{"type":"integer","format":"int32"},"deliveryTime":{"type":"string","format":"date-time"},"packAmount":{"type":"integer","format":"int32"},"tablewareNumber":{"type":"integer","format":"int32"},"tablewareStatus":{"type":"integer","format":"int32"},"orderDishes":{"type":"string","description":"订单菜品信息"},"orderDetailList":{"type":"array","description":"订单详情","items":{"$ref":"#/components/schemas/OrderDetail"}}},"description":"订单信息返回的数据格式"},"ResultOrderVO":{"type":"object","properties":{"code":{"type":"integer","description":"编码：1成功，0和其它数字为失败","format":"int32"},"msg":{"type":"string","description":"错误信息"},"data":{"$ref":"#/components/schemas/OrderVO"}},"description":"后端统一返回结果"},"PageResult":{"type":"object","properties":{"total":{"type":"integer","description":"总记录数","format":"int64"},"records":{"type":"array","description":"当前页数据集合","items":{"type":"object","description":"当前页数据集合"}}},"description":"分页查询结果"},"ResultPageResult":{"type":"object","properties":{"code":{"type":"integer","description":"编码：1成功，0和其它数字为失败","format":"int32"},"msg":{"type":"string","description":"错误信息"},"data":{"$ref":"#/components/schemas/PageResult"}},"description":"后端统一返回结果"},"DishFlavor":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"dishId":{"type":"integer","format":"int64"},"name":{"type":"string"},"value":{"type":"string"}},"description":"菜品关联的口味"},"DishVO":{"type":"object","properties":{"id":{"type":"integer","description":"菜品id","format":"int64"},"name":{"type":"string","description":"菜品名称"},"categoryId":{"type":"integer","description":"菜品分类id","format":"int64"},"price":{"type":"number","description":"菜品价格"},"image":{"type":"string","description":"图片"},"description":{"type":"string","description":"描述信息"},"status":{"type":"integer","description":"0 停售 1 起售","format":"int32"},"updateTime":{"type":"string","description":"更新时间","format":"date-time"},"categoryName":{"type":"string","description":"分类名称"},"flavors":{"type":"array","description":"菜品关联的口味","items":{"$ref":"#/components/schemas/DishFlavor"}}},"description":"菜品信息返回的数据格式"},"ResultListDishVO":{"type":"object","properties":{"code":{"type":"integer","description":"编码：1成功，0和其它数字为失败","format":"int32"},"msg":{"type":"string","description":"错误信息"},"data":{"type":"array","description":"数据","items":{"$ref":"#/components/schemas/DishVO"}}},"description":"后端统一返回结果"},"Category":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"type":{"type":"integer","format":"int32"},"name":{"type":"string"},"sort":{"type":"integer","format":"int32"},"status":{"type":"integer","format":"int32"},"createTime":{"type":"string","format":"date-time"},"updateTime":{"type":"string","format":"date-time"},"createUser":{"type":"integer","format":"int64"},"updateUser":{"type":"integer","format":"int64"}},"description":"数据"},"ResultListCategory":{"type":"object","properties":{"code":{"type":"integer","description":"编码：1成功，0和其它数字为失败","format":"int32"},"msg":{"type":"string","description":"错误信息"},"data":{"type":"array","description":"数据","items":{"$ref":"#/components/schemas/Category"}}},"description":"后端统一返回结果"},"ResultAddressBook":{"type":"object","properties":{"code":{"type":"integer","description":"编码：1成功，0和其它数字为失败","format":"int32"},"msg":{"type":"string","description":"错误信息"},"data":{"$ref":"#/components/schemas/AddressBook"}},"description":"后端统一返回结果"},"ResultListAddressBook":{"type":"object","properties":{"code":{"type":"integer","description":"编码：1成功，0和其它数字为失败","format":"int32"},"msg":{"type":"string","description":"错误信息"},"data":{"type":"array","description":"数据","items":{"$ref":"#/components/schemas/AddressBook"}}},"description":"后端统一返回结果"}},"securitySchemes":{"adminToken":{"type":"apiKey","name":"token","in":"header"},"userToken":{"type":"apiKey","name":"authentication","in":"header"}}}}